fix: explicit fixture building
