+++ .cargo/config.toml
@@ -1,0 +1,2 @@
+[alias]
+x = "run --package xtask --"
